Refactor shared functions
